feat(Core/Misc): moved Position to its own file
